+++ Sources/KeyboardKit/Input/InputSetRows.swift
@@ -1,5 +1,5 @@
 //
-//  KeyboardInputRows.swift
+//  InputSetRows.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-02-03.
@@ -11,16 +11,16 @@ import Foundation
 /**
  This typealias represents a list of keyboard input rows.
  */
-public typealias KeyboardInputRows = [KeyboardInputRow]
+public typealias InputSetRows = [InputSetRow]
 
-public extension KeyboardInputRows {
+public extension InputSetRows {
     
     /**
      Create input rows from a list of string array, that are
      mapped to a `KeyboardInputRow` array.
      */
     init(_ rows: [[String]]) {
-        self = rows.map { KeyboardInputRow($0) }
+        self = rows.map { InputSetRow($0) }
     }
     
     /**
@@ -32,7 +32,7 @@ public extension KeyboardInputRows {
     init(lowercased: [[String]], uppercased: [[String]]) {
         assert(lowercased.count == uppercased.count, "The lowercased and uppercased string arrays must contain the same amount of characters")
         self = lowercased.enumerated().map {
-            KeyboardInputRow(
+            InputSetRow(
                 lowercased: lowercased[$0.offset],
                 uppercased: uppercased[$0.offset])
         }