+++ Sources/KeyboardKit/_Deprecated/KeyboardInputRows.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+@available(*, deprecated, renamed: "InputSetRow")
+public typealias KeyboardInputRow = InputSetRow
+
+@available(*, deprecated, renamed: "InputSetRows")
+public typealias KeyboardInputRows = InputSetRows